// src/cfb/engine/types.ts
export type DefensiveFront = '4-2-5' | '4-3' | '3-3-5' | '3-4' | 'Multiple';
export type BaseCoverage = 'Cover-1' | 'Cover-2' | 'Cover-3' | 'Quarters' | 'Cover-2 Man';

export type OffensiveScheme =
  | 'Air Raid'
  | 'Spread'
  | 'RPO Spread'
  | 'Tempo Spread'
  | 'Pro Style'
  | 'Wide Zone'
  | 'Power Run'
  | 'Triple Option'
  | 'Vertical';